added liquibase, created new DTO for post operation, added manager filtration and new error for delete operation, small fixes

+++ src/main/java/com/employee_api/springboot_bootcamp/employee/EmployeeService.java
@@ -6,8 +6,10 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 import java.util.UUID;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 @Service
 @RequiredArgsConstructor
@@ -28,7 +30,24 @@ public class EmployeeService {
     }
 
     @Transactional
-    public void create(EmployeeDTO employeeDTO) {
+    public List<EmployeeDTO> getManagers(UUID id) {
+        List<EmployeeDTO> employees = this.getAll();
+        List<EmployeeDTO> subordinates = employees.stream().flatMap(employee -> {
+            if (employee.id().equals(id)) {
+                return Stream.of(employee);
+            }
+            if (Objects.nonNull(employee.manager()) && employee.manager().id().equals(id)) {
+                List<EmployeeDTO> subordinatesList = this.findSubordinates(employee, employees);
+                subordinatesList.add(employee);
+                return subordinatesList.stream();
+            }
+            return Stream.empty();
+        }).toList();
+        return employees.stream().filter(employee -> !subordinates.contains(employee)).toList();
+    }
+
+    @Transactional
+    public void create(PostEmployeeDTO employeeDTO) {
         employeeRepository.save(employeeMapper.mapToEntity(employeeDTO));
     }
 
@@ -48,7 +67,23 @@ public class EmployeeService {
 
     @Transactional
     public void delete(EmployeeDTO employeeDTO) {
+        List<EmployeeDTO> employees = this.getAll();
+        List<EmployeeDTO> subordinates = employees.stream()
+                .filter(employee -> Objects.nonNull(employee.manager()) && employee.manager().id().equals(employeeDTO.id())).toList();
+        if (!subordinates.isEmpty()) {
+            throw new IllegalArgumentException(String.format("Cannot delete %s %s because of manager position", employeeDTO.name(), employeeDTO.surname()));
+        }
         employeeRepository.delete(employeeMapper.mapToEntity(employeeDTO));
     }
 
+    private List<EmployeeDTO> findSubordinates(EmployeeDTO employeeDTO, List<EmployeeDTO> allEmployees) {
+        return allEmployees.stream().flatMap(employee -> {
+            if (Objects.nonNull(employee.manager()) && employee.manager().id().equals(employeeDTO.id())) {
+                List<EmployeeDTO> subordinatesList = this.findSubordinates(employee, allEmployees);
+                subordinatesList.add(employee);
+                return subordinatesList.stream();
+            }
+            return Stream.empty();
+        }).collect(Collectors.toList());
+    }
 }
